Moved render logic into search section component
